+++ cyclos_configuration/scripts/emandates/Library.properties
@@ -16,3 +16,5 @@ status.expired=De machtiging is verlopen zonder te zijn geautoriseerd. Maak alst
 status.cancelled=De machtiging is geannuleerd. Maak alstublieft een nieuwe aan.
 
 details=<br>Hieronder staan de details van uw huidige incassomachtiging:<br>&nbsp;
+
+errorSaveIBAN=Er is iets misgegaan bij het verwerken van uw machtiging. Neem alstublieft contact op met de administratie en geef daarbij de volgende melding door:
